Match existing hook without payload secret
The exsiting hook matcher started failing for the scenario where the
webhook does not define a payload_secret. It's content was set to
undefined, and the _.isMatch function used to filter the hooks was
returning undefined, as it didn't match any of the retrieved hooks,
even when it should have. This resulted in an attempt to create the
webhook, even if it exsited; github returned a 422 error.

This changes the desired config, to not include the secret property
if its value is undefined. With this change, the _.isMatch function
is able to find the match, and correctly updates or no-ops the hook.

The error returned by the resource before this changes looked like
the following:
```
Error while calling Github: StatusCodeError
Response Status: 422
Message: {
  "message": "Validation Failed",
  "errors": [
    {
      "resource": "Hook",
      "code": "custom",
      "message": "Hook already exists on this repository"
    },
    {
      "resource": "Hook",
      "code": "custom",
      "message": "Invalid config attribute: content-type"
    }
  ],
  "documentation_url": "https://docs.github.com/rest/repos/webhooks#create-a-repository-webhook",
  "status": "422"
}
```